requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
chromadb
tiktoken
pypdf
numpy
reportlab
beautifulsoup4
requests
